+++ modules/src/ics04_channel/handler/chan_open_confirm.rs
@@ -99,6 +99,7 @@ pub(crate) fn process(
 #[cfg(test)]
 mod tests {
     use std::convert::TryFrom;
+    use std::time::Duration;
 
     use crate::events::IbcEvent;
     use crate::ics02_client::client_type::ClientType;
@@ -137,7 +138,7 @@ mod tests {
             client_id.clone(),
             ConnectionCounterparty::try_from(get_dummy_raw_counterparty()).unwrap(),
             get_compatible_versions(),
-            0,
+            Duration::from_secs(0),
         );
 
         let msg_chan_confirm = MsgChannelOpenConfirm::try_from(