+++ relayer/src/config.rs
@@ -8,6 +8,7 @@ use tendermint_light_client::types::TrustThreshold;
 use ibc::ics24_host::identifier::{ChainId, PortId};
 
 use crate::error;
+use ibc::ics04_channel::channel::Order;
 
 /// Defaults for various fields
 pub mod default {
@@ -22,7 +23,15 @@ pub mod default {
     }
 
     pub fn clock_drift() -> Duration {
-        Duration::from_secs(5) // 5 seconds
+        Duration::from_secs(5)
+    }
+
+    pub fn connection_delay() -> Duration {
+        Duration::from_secs(0)
+    }
+
+    pub fn channel_ordering() -> Order {
+        Order::Unordered
     }
 }
 
@@ -43,16 +52,24 @@ impl Config {
     pub fn find_chain_mut(&mut self, id: &ChainId) -> Option<&mut ChainConfig> {
         self.chains.iter_mut().find(|c| c.id == *id)
     }
-    pub fn relay_paths(&self, src_chain: &ChainId, dst_chain: &ChainId) -> Option<Vec<RelayPath>> {
-        self.connections
-            .as_ref()?
-            .iter()
-            .find(|c| {
-                c.a_chain == *src_chain && c.b_chain == *dst_chain
-                    || c.a_chain == *dst_chain && c.b_chain == *src_chain
-            })?
-            .paths
-            .clone()
+
+    pub fn first_matching_path(
+        &self,
+        src_chain: &ChainId,
+        dst_chain: &ChainId,
+    ) -> Option<(&Connection, &RelayPath)> {
+        let connection = self.connections.as_ref()?.iter().find(|c| {
+            c.a_chain == *src_chain && c.b_chain == *dst_chain
+                || c.a_chain == *dst_chain && c.b_chain == *src_chain
+        });
+
+        connection.and_then(|conn| {
+            if let Some(ref paths) = conn.paths {
+                Some((conn, &paths[0]))
+            } else {
+                None
+            }
+        })
     }
 }
 
@@ -115,6 +132,8 @@ pub struct ChainConfig {
 pub struct Connection {
     pub a_chain: ChainId,
     pub b_chain: ChainId,
+    #[serde(default = "default::connection_delay", with = "humantime_serde")]
+    pub delay: Duration,
     pub paths: Option<Vec<RelayPath>>,
 }
 
@@ -122,6 +141,8 @@ pub struct Connection {
 pub struct RelayPath {
     pub a_port: PortId,
     pub b_port: PortId,
+    #[serde(default = "default::channel_ordering")]
+    pub ordering: Order,
 }
 
 /// Attempt to load and parse the TOML config file as a `Config`.