+++ relayer/src/relay.rs
@@ -1,3 +1,5 @@
+use std::time::Duration;
+
 use anomaly::BoxError;
 use tracing::info;
 
@@ -10,17 +12,18 @@ use crate::connection::Connection;
 use crate::foreign_client::ForeignClient;
 use crate::link::{Link, LinkParameters};
 
-pub(crate) const MAX_ITER: u32 = 10;
+pub(crate) const MAX_ITER: usize = 10;
 
 /// Used by the `hermes start ibc-0 ibc-1`
 pub fn relay_on_new_link(
     a_chain_handle: Box<dyn ChainHandle>,
     b_chain_handle: Box<dyn ChainHandle>,
+    delay: Duration,
     ordering: Order,
     path: RelayPath,
 ) -> Result<(), BoxError> {
     // Setup the clients, connection and channel
-    let channel = connect_with_new_channel(a_chain_handle, b_chain_handle, ordering, path)?;
+    let channel = connect_with_new_channel(a_chain_handle, b_chain_handle, delay, ordering, path)?;
 
     let mut link = Link::new(channel);
     link.relay()?;
@@ -29,7 +32,7 @@ pub fn relay_on_new_link(
 }
 
 /// Relays packets over a specified channel
-/// Used by the `hermes start ibc-0 ibc-1 transfer channel-0`
+/// Used by the `hermes start ibc-0 ibc-1 --src-port-id transfer --src-channel-id channel-0`
 pub fn channel_relay(
     a_chain: Box<dyn ChainHandle>,
     b_chain: Box<dyn ChainHandle>,
@@ -40,10 +43,11 @@ pub fn channel_relay(
 }
 
 /// Connects two ports of two chains creating new clients, connection and channel
-/// Used by the `hermes channel handshake ibc-0 ibc-1 transfer transfer `
+/// Used by the `hermes channel handshake ibc-0 ibc-1`
 pub fn connect_with_new_channel(
     a_chain_handle: Box<dyn ChainHandle>,
     b_chain_handle: Box<dyn ChainHandle>,
+    delay: Duration,
     ordering: Order,
     path: RelayPath,
 ) -> Result<Channel, BoxError> {
@@ -54,7 +58,7 @@ pub fn connect_with_new_channel(
     let client_on_b = ForeignClient::new(b_chain_handle.clone(), a_chain_handle.clone())?;
 
     // Setup the connection between the two chains
-    let connection = Connection::new(client_on_a, client_on_b, 0)?;
+    let connection = Connection::new(client_on_a, client_on_b, delay)?;
 
     // Setup the channel over the connection
     Ok(Channel::new(